WIP: Implement advance-branches for jj commit
## TODO
There is still a lot of work to do for this feature. We need to support `jj
new`, and we need to add a `--branch` argument to `jj new` (and possibly `jj
commit`) to resolve instances where more than one branch would be moved. The
--branch option for `jj new` will allow users to create a new topic branch on
top of an existing branch such as `main` without moving the `main` branch or
needing to set an override for it in the config.toml.

## Feature Description

If enabled in the user or repository settings, the local branch pointing to the
revision targeted by `jj commit` will be advanced to the newly created commit.

This behavior can be enabled by default for all branches by setting
the following in the config.toml:

```
[branches]
advance-branches = true
```

The default value of `advance-branches` is `false`. You can override the
behavior for specific branches by setting the value of
`advance-branches-overrides` in the config.toml:

```
[branches]
advance-branches-overrides = ["main"]
```

Branches that have been overridden behave as if the value of `advance-branches`
was negated.


This implements feature request 2338.