SQUASH: Advance all eligible branches at the same time instead of a single branch
If more than one branch is eligible to advance, advance them all instead of
complaining about ambiguity and returning an error to the user. There's no
fundamental reason why we cannot or should not advance all eligible branches.

It is posible for users to get into a situation where they have no easy way to
adjust the repo such that only one branch is eligible to advance. This is
because `jj new` will also try to advance branches (once implemented), meaning
there's no easy way to create a new commit and move the target branch to it
without adding a flag to `jj new` to disable the advanceable branch behavior.

We want this feature to be controlled by the config only, so adding a flag to
`jj new` is not an option.

Simply advancing all eligible branches avoid difficult-to-resolve ambiguous
situations. The only downside is that for colocated repos we will have to
detach the Git HEAD instead of checking out a single branch if multiple
branches point to @-. I think this is an acceptable compromise to make `jj new`
and `jj commit` more ergonomic when this feature is enabled.